requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
mssql-django>=1.5
pyodbc>=5.1
python-dotenv>=1.0
numpy>=1.26
scipy>=1.11
djangorestframework-simplejwt==5.5.1
